Docker

+++ ServicioAdministracion/src/main/java/utnfc/isi/back/sim/repository/CamionRepository.java
@@ -3,6 +3,6 @@ package utnfc.isi.back.sim.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import utnfc.isi.back.sim.domain.Camion;
 
-public interface CamionRepository extends JpaRepository<Camion, Integer> {
+public interface CamionRepository extends JpaRepository<Camion, Long> {
     Camion findByDominio(String dominio);
 }
