+++ ServicioAdministracion/src/main/java/utnfc/isi/back/sim/repository/TarifaRepository.java
@@ -3,5 +3,5 @@ package utnfc.isi.back.sim.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import utnfc.isi.back.sim.domain.Tarifa;
 
-public interface TarifaRepository extends JpaRepository<Tarifa, Integer> {
+public interface TarifaRepository extends JpaRepository<Tarifa, Long> {
 }
